ast refactor import cleanup
