+++ src/main/java/fr/inria/corese/core/next/query/api/io/parser/QueryParser.java
@@ -1,6 +1,6 @@
 package fr.inria.corese.core.next.query.api.io.parser;
 
-import fr.inria.corese.core.next.query.api.sparql.ast.QueryAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.QueryAst;
 
 import java.io.InputStream;
 import java.io.Reader;
