--- src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/ConstructQueryAst.java
@@ -1,10 +1,0 @@
-package fr.inria.corese.core.next.query.api.sparql.ast;
-
-/**
- * AST representation of a SPARQL CONSTRUCT query.
- */
-public non-sealed interface ConstructQueryAst extends QueryAst {
-    GroupGraphPatternAst constructTemplate();
-    @Override
-    GroupGraphPatternAst whereClause();
-}
