--- src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/IriAst.java
@@ -1,12 +1,0 @@
-package fr.inria.corese.core.next.query.api.sparql.ast;
-
-/**
- * IRI or QName in a triple pattern (e.g. &lt;http://...&gt;, foaf:Person, a).
- */
-public record IriAst(String raw) implements TermAst {
-    public IriAst {
-        if (raw == null) {
-            throw new IllegalArgumentException("IRI raw is null");
-        }
-    }
-}
