--- src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/PatternAst.java
@@ -1,7 +1,0 @@
-package fr.inria.corese.core.next.query.api.sparql.ast;
-
-/**
- * Element of a group graph pattern (BGP, optional, union, etc.).
- */
-public sealed interface PatternAst permits BgpAst {
-}
