--- src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/SelectQueryAst.java
@@ -1,14 +1,0 @@
-package fr.inria.corese.core.next.query.api.sparql.ast;
-
-import java.util.List;
-
-/**
- * Abstract Syntax Tree (AST) representation of a SPARQL {@code SELECT} query.
- */
-public record SelectQueryAst(GroupGraphPatternAst whereClause) implements QueryAst {
-    public SelectQueryAst {
-        if (whereClause == null) {
-            whereClause = new GroupGraphPatternAst(List.of());
-        }
-    }
-}
