--- src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/TermAst.java
@@ -1,7 +1,0 @@
-package fr.inria.corese.core.next.query.api.sparql.ast;
-
-/**
- * AST node for a SPARQL term (variable, IRI or literal) in a triple pattern.
- */
-public sealed interface TermAst permits VarAst, IriAst, LiteralAst {
-}
