--- src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/VarAst.java
@@ -1,12 +1,0 @@
-package fr.inria.corese.core.next.query.api.sparql.ast;
-
-/**
- * Variable in a triple pattern (e.g. ?s, $x).
- */
-public record VarAst(String name) implements TermAst {
-    public VarAst {
-        if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("Variable name is null or blank");
-        }
-    }
-}
