+++ src/main/java/fr/inria/corese/core/next/query/impl/parser/SparqlParser.java
@@ -1,13 +1,15 @@
 package fr.inria.corese.core.next.query.impl.parser;
 
+import fr.inria.corese.core.next.data.impl.io.common.IOConstants;
 import fr.inria.corese.core.next.impl.parser.antlr.SparqlLexer;
 import fr.inria.corese.core.next.query.api.exception.QueryException;
 import fr.inria.corese.core.next.query.api.exception.QuerySyntaxException;
 import fr.inria.corese.core.next.data.impl.io.parser.util.ParserConstants;
 import fr.inria.corese.core.next.query.api.base.io.AbstractQueryParser;
 import fr.inria.corese.core.next.query.api.io.parser.QueryOptions;
-import fr.inria.corese.core.next.query.api.sparql.ast.QueryAst;
-import fr.inria.corese.core.next.query.impl.parser.listener.SparqlBgpFeature;
+import fr.inria.corese.core.next.query.impl.sparql.ast.QueryAst;
+import fr.inria.corese.core.next.query.impl.parser.listener.BgpFeature;
+import fr.inria.corese.core.next.query.impl.sparql.options.BaseIRIOptions;
 import org.antlr.v4.runtime.*;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
@@ -31,7 +33,11 @@ public class SparqlParser extends AbstractQueryParser {
     }
     @Override
     public QueryAst parse(InputStream in) {
-        return parse(new java.io.InputStreamReader(in, StandardCharsets.UTF_8), null);
+        String baseIri = IOConstants.getDefaultBaseURI();
+        if (getConfig() instanceof BaseIRIOptions baseIRIOptions) {
+            baseIri = baseIRIOptions.getBaseIRI();
+        }
+        return parse(new java.io.InputStreamReader(in, StandardCharsets.UTF_8), baseIri);
     }
 
     @Override
@@ -95,7 +101,7 @@ public class SparqlParser extends AbstractQueryParser {
             SparqlAstBuilder builder = new SparqlAstBuilder(sparqlParserOptions);
 
             SparqlListener listener = new SparqlListener(List.of(
-                    new SparqlBgpFeature(builder)
+                    new BgpFeature(builder)
             ));
 
             walker.walk(listener, tree);
