+++ src/main/java/fr/inria/corese/core/next/query/impl/parser/listener/AbstractSparqlFeature.java
@@ -1,4 +1,4 @@
-package fr.inria.corese.core.next.query.sparql.parser.listener;
+package fr.inria.corese.core.next.query.impl.parser.listener;
 
 import fr.inria.corese.core.next.impl.parser.antlr.SparqlParserBaseListener;
