+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/BgpAst.java
@@ -5,7 +5,7 @@ import java.util.List;
 /**
  * Basic Graph Pattern: a list of triple patterns (TriplesBlock).
  */
-public record BgpAst(List<TriplePatternAst> triples) {
+public record BgpAst(List<TriplePatternAst> triples) implements PatternAst {
     public BgpAst {
         triples = triples != null ? List.copyOf(triples) : List.of();
     }
