+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/LiteralAst.java
@@ -3,7 +3,7 @@ package fr.inria.corese.core.next.query.impl.sparql.ast;
 /**
  * RDF literal in a triple pattern (lexical form, optional language tag or datatype).
  */
-public record LiteralAst(String lexical, String lang, String datatype) {
+public record LiteralAst(String lexical, String lang, String datatype) implements TermAst {
     public LiteralAst {
         if (lexical == null) {
             throw new IllegalArgumentException("Literal lexical is null");
