+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/PatternAst.java
@@ -3,5 +3,5 @@ package fr.inria.corese.core.next.query.impl.sparql.ast;
 /**
  * Element of a group graph pattern (BGP, optional, union, etc.).
  */
-public interface PatternAst {
+public sealed interface PatternAst permits BgpAst {
 }
