+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/QueryAst.java
@@ -1,12 +1,9 @@
 package fr.inria.corese.core.next.query.impl.sparql.ast;
 
-import fr.inria.corese.core.next.query.sparql.ast.*;
-
-
 /**
  * Minimal SPARQL query AST (SELECT, ASK, CONSTRUCT, DESCRIBE).
  * Holds the WHERE clause as a group graph pattern; query-specific projection/template can be added later.
  */
-public sealed interface QueryAst permits SelectQueryAst {
+public sealed interface QueryAst permits AskQueryAst, ConstructQueryAst, DescribeQueryAst, SelectQueryAst {
     GroupGraphPatternAst whereClause();
 }
