+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/TermAst.java
@@ -3,5 +3,7 @@ package fr.inria.corese.core.next.query.impl.sparql.ast;
 /**
  * AST node for a SPARQL term (variable, IRI or literal) in a triple pattern.
  */
-public interface TermAst {
+public sealed interface TermAst permits VarAst, IriAst, LiteralAst {
 }
+
+
