+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/TriplePatternAst.java
@@ -1,7 +1,5 @@
 package fr.inria.corese.core.next.query.impl.sparql.ast;
 
-import fr.inria.corese.core.next.query.api.sparql.ast.TermAst;
-
 /**
  * A single triple pattern (s p o) in a BGP.
  */
