+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/VarAst.java
@@ -3,7 +3,7 @@ package fr.inria.corese.core.next.query.impl.sparql.ast;
 /**
  * Variable in a triple pattern (e.g. ?s, $x).
  */
-public record VarAst(String name) {
+public record VarAst(String name) implements TermAst {
     public VarAst {
         if (name == null || name.isBlank()) {
             throw new IllegalArgumentException("Variable name is null or blank");
