+++ src/test/java/fr/inria/corese/core/next/query/api/sparql/ast/SparqlAstTest.java
@@ -1,5 +1,6 @@
 package fr.inria.corese.core.next.query.api.sparql.ast;
 
+import fr.inria.corese.core.next.query.impl.sparql.ast.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
