+++ src/test/java/fr/inria/corese/core/next/query/impl/parser/SparqlListenerTest.java
@@ -1,11 +1,11 @@
 package fr.inria.corese.core.next.query.impl.parser;
 
 import fr.inria.corese.core.next.impl.parser.antlr.SparqlLexer;
-import fr.inria.corese.core.next.query.api.sparql.ast.BgpAst;
-import fr.inria.corese.core.next.query.api.sparql.ast.QueryAst;
-import fr.inria.corese.core.next.query.api.sparql.ast.TriplePatternAst;
-import fr.inria.corese.core.next.query.api.sparql.ast.VarAst;
-import fr.inria.corese.core.next.query.impl.parser.listener.SparqlBgpFeature;
+import fr.inria.corese.core.next.query.impl.sparql.ast.BgpAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.QueryAst;
+import fr.inria.corese.core.next.query.impl.parser.listener.BgpFeature;
+import fr.inria.corese.core.next.query.impl.sparql.ast.TriplePatternAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.VarAst;
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
@@ -38,9 +38,9 @@ class SparqlListenerTest {
 
     @Test
     void constructorWithDelegatesCopiesList() {
-        SparqlBgpFeature feature = new SparqlBgpFeature(
+        BgpFeature feature = new BgpFeature(
                 new SparqlAstBuilder(new SparqlParserOptions.Builder().build()));
-        List<SparqlBgpFeature> mutable = new java.util.ArrayList<>(List.of(feature));
+        List<BgpFeature> mutable = new java.util.ArrayList<>(List.of(feature));
         SparqlListener listener = new SparqlListener(mutable);
         mutable.clear();
         // Listener should still have one delegate (defensive copy)
@@ -55,7 +55,7 @@ class SparqlListenerTest {
     void withSingleBgpDelegateWalkProducesAst() {
         SparqlParserOptions opts = new SparqlParserOptions.Builder().build();
         SparqlAstBuilder builder = new SparqlAstBuilder(opts);
-        SparqlListener listener = new SparqlListener(List.of(new SparqlBgpFeature(builder)));
+        SparqlListener listener = new SparqlListener(List.of(new BgpFeature(builder)));
 
         fr.inria.corese.core.next.impl.parser.antlr.SparqlParser antlrParser = createAntlrParser("SELECT * WHERE { ?s ?p ?o }");
         new ParseTreeWalker().walk(listener, antlrParser.query());
