+++ src/test/java/fr/inria/corese/core/next/query/sparql/parser/SparqlListenerTest.java
@@ -1,6 +1,14 @@
 package fr.inria.corese.core.next.query.sparql.parser;
 
 import fr.inria.corese.core.next.impl.parser.antlr.SparqlLexer;
+import fr.inria.corese.core.next.query.impl.parser.SparqlAstBuilder;
+import fr.inria.corese.core.next.query.impl.parser.SparqlListener;
+import fr.inria.corese.core.next.query.impl.parser.SparqlParserOptions;
+import fr.inria.corese.core.next.query.impl.parser.listener.BgpFeature;
+import fr.inria.corese.core.next.query.impl.sparql.ast.BgpAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.QueryAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.TriplePatternAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.VarAst;
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
@@ -33,9 +41,9 @@ class SparqlListenerTest {
 
     @Test
     void constructorWithDelegatesCopiesList() {
-        SparqlBgpFeature feature = new SparqlBgpFeature(
+        BgpFeature feature = new BgpFeature(
                 new SparqlAstBuilder(new SparqlParserOptions.Builder().build()));
-        List<SparqlBgpFeature> mutable = new java.util.ArrayList<>(List.of(feature));
+        List<BgpFeature> mutable = new java.util.ArrayList<>(List.of(feature));
         SparqlListener listener = new SparqlListener(mutable);
         mutable.clear();
         // Listener should still have one delegate (defensive copy)
@@ -50,7 +58,7 @@ class SparqlListenerTest {
     void withSingleBgpDelegateWalkProducesAst() {
         SparqlParserOptions opts = new SparqlParserOptions.Builder().build();
         SparqlAstBuilder builder = new SparqlAstBuilder(opts);
-        SparqlListener listener = new SparqlListener(List.of(new SparqlBgpFeature(builder)));
+        SparqlListener listener = new SparqlListener(List.of(new BgpFeature(builder)));
 
         fr.inria.corese.core.next.impl.parser.antlr.SparqlParser antlrParser = createAntlrParser("SELECT * WHERE { ?s ?p ?o }");
         new ParseTreeWalker().walk(listener, antlrParser.query());
